REFACTOR: Break the tests by removing the chaining code
Normally you would not leave a build broken, but I commit it now to show that it can help detect
where certain code is used. And where you need to go in order to fix it. A good test set can really
guide you with that, without fear of breaking other functionality